Fix empty recipe list crash
Recipe list view was crashing when the api call that populates drinks
found nothing. So now it returns an empty array instead of causing a
fatal error. The view now checks when that array is empty and if it is, displays a message prompting the user to select some bottles or change their settings.