Make it possible to remove items from homebar list
This commit disables the edit button when ingredients list is empty, and
when ingredients list is not empty changes the edit button to offer
deletion only. It also adds a "Add more bottles" button to separate
adding/ removing bottles. (This is a workaround because it isn't
possible to see previous selections in edit mode)