Set up preferred bottle selection for recipe list
This commit adds a selectedBottle property to RecipeSettingsView,
HomeBar, and the CocktailAPI. Once a bottle is selected using the picker
in RecipeSettingsView, that value is used to update the selectedBottle
property of HomeBar so that the recipe list can base it's fetch on that
selection. CocktailAPI was also refactored to allow for this, in
particular setting the buildDrinkStubListUrl method to be based on just
the selectedBottle rather than the whole bottleList.