Add ingredients to HomeBar and fix bug
This commit fixes the bug where the IngredientView ingredients list
flickers and wont let the user select anything. This is caused by the
list refreshing on every tap (SwiftUI). Instead, it moves the code that
checks for the network availability (by proxy, using the API call) to
the HomeBar class and has HomeBar create the ingredients list,
using either the result of the API call or the saved list.